CAMEL-16280: Extract EndpointUriFactory test support to camel-test-junit5
