+++ components/camel-netty-http/src/test/java/org/apache/camel/component/netty/http/NettyHttpEndpointUriFactoryTest.java
@@ -19,41 +19,23 @@ package org.apache.camel.component.netty.http;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.apache.camel.ExtendedCamelContext;
-import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.spi.EndpointUriFactory;
 import org.apache.camel.test.AvailablePortFinder;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.Test;
+import org.apache.camel.test.junit5.EndpointUriFactoryTestSupport;
 
-public class NettyHttpEndpointUriFactoryTest extends CamelTestSupport {
+public class NettyHttpEndpointUriFactoryTest extends EndpointUriFactoryTestSupport {
 
-    @Test
-    void buildUri() {
-        // let's just see if the route starts successfully
-    }
-
-    private String scheme() {
+    @Override
+    protected String scheme() {
         return "netty-http";
     }
 
-    private Map<String, Object> pathParameters() {
+    @Override
+    protected Map<String, Object> pathParameters() {
         Map<String, Object> params = new HashMap<>();
         params.put("protocol", "http");
         params.put("host", "localhost");
         params.put("port", AvailablePortFinder.getNextAvailable());
         params.put("path", "test");
         return params;
-    }
-
-    @Override
-    protected RouteBuilder createRouteBuilder() {
-        return new RouteBuilder() {
-            public void configure() throws Exception {
-                EndpointUriFactory factory = getContext().adapt(ExtendedCamelContext.class).getEndpointUriFactory(scheme());
-                String uri = factory.buildUri(scheme(), pathParameters(), false);
-                from(uri).to("mock:out");
-            }
-        };
     }
 }
